refactor: consolidate `run-monitoring.sh` output files

+++ qa-physics/datasetOrganize.sh
@@ -5,9 +5,14 @@ set -e
 if [ $# -ne 1 ];then echo "USAGE: $0 [dataset]" >&2; exit 101; fi
 dataset=$1
 
+if [ -z "$CLASQA" ]; then
+  echo "ERROR: please source environ.sh first" >&2
+  exit 100
+fi
+
 # cleanup / generate new dataset subdirs
 for outdir in outmon outdat; do
-  dir=${outdir}.${dataset}
+  dir=$CLASQA/${outdir}.${dataset}
   echo "clean $dir"
   mkdir -p $dir
   rm -r $dir
@@ -16,13 +21,14 @@ done
 
 # loop over runs, copying and linking to dataset subdirs
 source datasetListParser.sh $dataset
-for file in outmon/monitor_*.hipo; do
+INPUT_DIR=$(realpath $CLASQA/../outfiles/physics)
+for file in $INPUT_DIR/monitor_*.hipo; do
   run=$(echo $file | sed 's/^.*monitor_//'|sed 's/\.hipo$//')
 
   if [ $run -ge $RUNL -a $run -le $RUNH ]; then
     echo "file run $run to dataset $dataset"
-    cat outdat/data_table_${run}.dat >> outdat.${dataset}/data_table.dat
-    ln -sv `pwd`/outmon/monitor_${run}.hipo ./outmon.${dataset}/monitor_${run}.hipo
+    cat $INPUT_DIR/data_table_${run}.dat >> $CLASQA/outdat.${dataset}/data_table.dat
+    ln -sv $INPUT_DIR/monitor_${run}.hipo $CLASQA/outmon.${dataset}/monitor_${run}.hipo
   fi
 
 done
